+++ middleware/logger.go
@@ -10,8 +10,12 @@ import (
 )
 
 var (
+	// LogEntryCtxKey is the context.Context key to store the request log entry.
 	LogEntryCtxKey = &contextKey{"LogEntry"}
 
+	// DefaultLogger is called by the Logger middleware handler to log each request.
+	// Its made a package-level variable so that it can be reconfigured for custom
+	// logging configurations.
 	DefaultLogger = RequestLogger(&DefaultLogFormatter{Logger: log.New(os.Stdout, "", log.LstdFlags)})
 )
 
@@ -27,6 +31,7 @@ func Logger(next http.Handler) http.Handler {
 	return DefaultLogger(next)
 }
 
+// RequestLogger returns a logger handler using a custom LogFormatter.
 func RequestLogger(f LogFormatter) func(next http.Handler) http.Handler {
 	return func(next http.Handler) http.Handler {
 		fn := func(w http.ResponseWriter, r *http.Request) {
@@ -44,29 +49,37 @@ func RequestLogger(f LogFormatter) func(next http.Handler) http.Handler {
 	}
 }
 
+// LogFormatter initiates the beginning of a new LogEntry per request.
+// See DefaultLogFormatter for an example implementation.
 type LogFormatter interface {
 	NewLogEntry(r *http.Request) LogEntry
 }
 
+// LogEntry records the final log when a request completes.
+// See defaultLogEntry for an example implementation.
 type LogEntry interface {
 	Write(status, bytes int, elapsed time.Duration)
 	Panic(v interface{}, stack []byte)
 }
 
+// GetLogEntry returns the in-context LogEntry for a request.
 func GetLogEntry(r *http.Request) LogEntry {
 	entry, _ := r.Context().Value(LogEntryCtxKey).(LogEntry)
 	return entry
 }
 
+// WithLogEntry sets the in-context LogEntry for a request.
 func WithLogEntry(r *http.Request, entry LogEntry) *http.Request {
 	r = r.WithContext(context.WithValue(r.Context(), LogEntryCtxKey, entry))
 	return r
 }
 
+// DefaultLogFormatter is a simple logger that implements a LogFormatter.
 type DefaultLogFormatter struct {
 	Logger *log.Logger
 }
 
+// NewLogEntry creates a new LogEntry for the request.
 func (l *DefaultLogFormatter) NewLogEntry(r *http.Request) LogEntry {
 	entry := &defaultLogEntry{
 		DefaultLogFormatter: l,